Move waker package to top level of internal packages.
