+++ internal/tailer/logstream/filestream_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/golang/glog"
 	"github.com/google/mtail/internal/logline"
 	"github.com/google/mtail/internal/tailer/logstream"
-	"github.com/google/mtail/internal/tailer/waker"
 	"github.com/google/mtail/internal/testutil"
+	"github.com/google/mtail/internal/waker"
 )
 
 func TestFileStreamRead(t *testing.T) {
