+++ internal/tailer/logstream/logstream.go
@@ -16,7 +16,7 @@ import (
 	"time"
 
 	"github.com/google/mtail/internal/logline"
-	"github.com/google/mtail/internal/tailer/waker"
+	"github.com/google/mtail/internal/waker"
 )
 
 // LogStream
